+++ src/main/java/com/github/idea/ginkgo/GinkgoTestEventsConverterV2.java
@@ -4,7 +4,7 @@ import com.intellij.execution.testframework.TestConsoleProperties;
 import com.intellij.openapi.util.Key;
 import jetbrains.buildServer.messages.serviceMessages.ServiceMessageVisitor;
 import jetbrains.buildServer.messages.serviceMessages.TestStdErr;
-import org.codehaus.plexus.util.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 
 import java.text.ParseException;
@@ -44,6 +44,7 @@ public class GinkgoTestEventsConverterV2 extends GinkgoTestEventsConverter {
             suites.push(suiteName);
             startTest(suiteName, outputType, visitor);
             processOutput(line, outputType, visitor);
+
             return line.length();
         }